serverをpresentation層に移動
